backend/requirements.txt
attrs==21.2.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
idna==2.10
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
packaging==20.9
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
PyYAML==5.4.1
requests==2.25.1
toml==0.10.2
urllib3==1.26.4
Werkzeug==1.0.1
